update: all features with verifytoken middleware

+++ routes/todos.js
@@ -6,19 +6,19 @@ const { verifyToken } = require('./middleware');
 const connection = mysql.createConnection(dbconfig);
 const router = express.Router();
 
-// router.post('/', verifyToken, (req, res) => {
-router.post('/', (req, res) => {
+router.post('/', verifyToken, (req, res) => {
+// router.post('/', (req, res) => {
   // console.log(req.decoded, 1);
   // console.log(req.body, 1);
   const { content, subjectId } = req.body;
-  // const { userInfo } = req.decoded;
+  const { userInfo } = req.decoded;
   // console.log(typeof parseInt(subjectId, 10));
   if (content?.length === 0 || content === undefined) {
     return res.json({ error: 'empty content field' });
   }
   try {
-    const insertsql = `INSERT INTO todos(content, subject_id, user_id) VALUES("${content}", ${parseInt(subjectId, 10)}, ${parseInt(1, 10)})`;
-    // const insertsql = `INSERT INTO todos(content, subject_id, user_id) VALUES("${content}", ${parseInt(subjectId, 10)}, ${userInfo.id})`;
+    // const insertsql = `INSERT INTO todos(content, subject_id, user_id) VALUES("${content}", ${parseInt(subjectId, 10)}, ${parseInt(1, 10)})`;
+    const insertsql = `INSERT INTO todos(content, subject_id, user_id) VALUES("${content}", ${parseInt(subjectId, 10)}, ${userInfo.id})`;
     connection.query(insertsql, (insertErr, result, fields) => {
       console.log(result);
       if (insertErr) throw insertErr;
@@ -27,8 +27,8 @@ router.post('/', (req, res) => {
           id: result.insertId,
           content,
           subject_id: parseInt(subjectId, 10),
-          // userId: userInfo.id,
-          user_id: 1,
+          userId: userInfo.id,
+          // user_id: 1,
           is_done: false,
         },
       });
@@ -40,11 +40,11 @@ router.post('/', (req, res) => {
 
 // todo 이미 있는 내용은 추가로 생성하지 않게 해야할까? 같은 것이 들어가면 안되는 이유도 없으니 그냥 유저가 알아서 지우게 하는게 나을 듯
 
-// router.patch('/:todoID', verifyToken, (req, res) => {
-router.patch('/:todoID', (req, res) => {
+router.patch('/:todoID', verifyToken, (req, res) => {
+// router.patch('/:todoID', (req, res) => {
   const { content, isDone } = req.body;
-  // const { userInfo } = req.decoded;
-  const userInfo = {id:1};
+  const { userInfo } = req.decoded;
+  // const userInfo = {id:1};
   const { todoID } = req.params;
   // console.log(todoId)
   try {
@@ -70,10 +70,12 @@ router.patch('/:todoID', (req, res) => {
   }
 });
 
-router.delete('/:todoID', (req, res) => {
+router.delete('/:todoID', verifyToken, (req, res) => {
   //const user_id = // 쿼리 수정
+  const user_id = req.decoded.userInfo.id;
+  console.log(req.decoded.userInfo.id, req.decoded.userInfo.username);
   const { todoID } = req.params;
-  connection.query(`DELETE FROM todos WHERE id = ${todoID};`, (err, result) => {
+  connection.query(`DELETE FROM todos WHERE id = ${todoID} AND user_id = ${user_id};`, (err, result) => {
     if (err) throw err;
     console.log(result);
     return res.status(200).send({ message: 'success' });
